Fixed PR feedback
